+++ app/config/dump/options.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"log/slog"
 
-	iiifconfig "github.com/go-iiif/go-iiif/v7/config"
+	iiifconfig "github.com/go-iiif/go-iiif/v8/config"
 	"github.com/sfomuseum/go-flags/flagset"
 )