+++ driver/driver.go
@@ -8,10 +8,10 @@ import (
 	"strings"
 
 	"github.com/aaronland/go-roster"
-	iiifcache "github.com/go-iiif/go-iiif/v7/cache"
-	iiifconfig "github.com/go-iiif/go-iiif/v7/config"
-	iiifimage "github.com/go-iiif/go-iiif/v7/image"
-	iiifsource "github.com/go-iiif/go-iiif/v7/source"
+	iiifcache "github.com/go-iiif/go-iiif/v8/cache"
+	iiifconfig "github.com/go-iiif/go-iiif/v8/config"
+	iiifimage "github.com/go-iiif/go-iiif/v8/image"
+	iiifsource "github.com/go-iiif/go-iiif/v8/source"
 )
 
 type Driver interface {